fix(session/context): resolve compilation errors from sub-package extraction
- Fix package declarations in test files (session_test → context_test, session → context)
- Fix append() closing paren (was } instead of ))
- Fix intra-package ContextStrategy/PipelineFactory references
- Fix Metadata import path (session → sessctx)
- Fix WithLogger/WithSessionProvider import paths
- Remove unused session imports from engine.go and orchestratortest

Refs #88
