+++ internal/agent/session/context/context_manager_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 gosharplite@gmail.com
 // SPDX-License-Identifier: MIT
 
-package session_test
+package context_test
 
 import (
 	"context"
