+++ internal/agent/session/context/warning_injector.go
@@ -13,7 +13,7 @@ import (
 
 // WarningInjector adds safety warnings to the context.
 type WarningInjector struct {
-	Strategy *ContextStrategy
+	Strategy *Strategy
 }
 
 func (t *WarningInjector) Transform(ctx context.Context, req *ports.ContextRequest) error {
